moved all pages to src/pages
