+++ codestrike frontend/src/pages/homepage.css
