+++ codestrike frontend/src/pages/logo.png
